Move longToPaddedHex function out of Tracers
